Move the overlay_action to utils
